Python: fix a problem with symbol versioning
The problem is described and explained here:

http://chris.dzombak.name/blog/2010/03/building-openssl-with-symbol-versioning.html

but the solution in our case is to help gcc to look at our own library first
(thanks to the -L flag), then it gets the right symbol versioning and there is
no problem. For some reason, without the -L flag, even though there is the
-Wl,-rpath flag, it still looks at Ubuntu package first. This commit fixes the
problem by adding the -L flag for openssl.